Add files via upload
This program is designed to assign tasks to different processors in a computer system, aiming to do so in the most efficient way possible. Here's how it works:

Setting Up the Processors and Tasks:

The program creates two types of "maps": one for processors (the devices that will perform the tasks) and another for the tasks themselves.
Each processor has specific capabilities, like how fast it can run (its CPU) and how much energy it uses. Each task has certain needs, like how much CPU power it requires to complete.
Calculating Costs:

Time: The program figures out how long it will take for each processor to complete its assigned tasks, based on the CPU needs of each task and the power of each processor.
Energy: It also calculates the energy each processor will use while working on its tasks.
Communication: If two tasks depend on each other but are assigned to different processors, there’s a "communication cost" involved in sending data between them.
Optimizing the Assignment:

The goal is to find the best way to assign tasks to processors so that the overall time, energy, and communication costs are as low as possible.
The program makes sure that no processor is given more tasks than it can handle, based on its CPU capacity.
Finding the Best Solution:

Using a method called optimization, the program tests different ways of assigning tasks to processors to find the most efficient setup.
In the end, it tells you which tasks should go to which processors and what the total cost (in time, energy, and communication) will be for this optimal setup.
In simple terms, the program is like a manager who figures out the best way to divide work among workers (processors) to get everything done as quickly and efficiently as possible, without overworking anyone.

The program is written in Google Colab